Rename image-attachment-picker.native.pure.ts to picked-image-normalizer.ts
