+++ packages/app/src/hooks/picked-image-normalizer.ts
